"Bug" fix - if more than 1 planets were placed, previous planets would stop orbiting and their state would all be messed up.
After some research, the fault was that I was handing pointers to Planet[0] from the main PlanetVector. But as its size grew, it reallocated its elements, leaving Planet[0] with an invalid pointer.

Implemented solution was to use new allocation of planets and use NewPointer-Handles to keep track of vector elements.